Polish UI and Add Custom Separators to Backend
UI now displays a summary of the current leaf for the user to view
before selecting to save the leaf.

Custom separators can now be used with the codes in the format string.
If the corresponding node is an array, the custom separator is used to
join the array elements into one string before inserting at the key's
position in the format string.

E.g. format string is 'Here is a list of parts: {part:\n}'
This would result in the array of parts being joined by \n into one
string to be inserted at the position of {part}