Refactor To Hard-Code Notes Section
The notes are now to be structured in the form [Opening Notes] [All
Doors] [Specific Door Sections] [Remaining Doors] [Closing Notes].

This is hard-coded and not related to the ui-tree anymore.

The specific door section is replaced by custom selections of serial
numbers so that notes can be assigned to specific doors only and listed
in their own section.

UI setup to navigate appropriately, backend does not yet reflect the
change to hard-code these note sections.