Add Search Page
Search processes the user's query as a regexp.

It first converts any spaces into .*? to ensure quick searching can be
done, but other than that it passes the user's string directly into new
RegExp() allowing full regexp to be used.

It compares the label of each node, concatenated with the labels of all
of it's parent nodes against the search term. In this way multiple terms
can be searched across multiple levels of the ui-tree.

Performance could be improved, also there is a bit of a big which causes
the user to have to press a tile twice before it registers as as
submission.