Add Delete Functionality to Leaf Tile Component
Also setup skeleton of the favourite functionality for the leaf tile.

Can now pass an onDelete or onFav callback to the leaftile which
operates on the path to the node of the tile.

Delete buttons are positioned in header and when pressed display a
confirmation box.

Styles of leaf tiles have been improved.