Add Review Component and Refactor Main Component
Refactored the sections code in the main component to be simpler and
allow easier integration with other components.

Created the leaf review component which displays the current leaves
using the LeafTile component and will soon allow the user to delete
them.

Styles need work and the delete functionality needs to be added.