Modify .gitignore to Work Better with CREDENTIALS
CREDENTIALS needs to be untracked to keep things private, but sometimes
needs to be tracked to be pushed to the prod server.

If tracked in the prod branch, the file is deleted when switching back
to dev (dev being published on github).

So a copy of CREDENTIALS has been placed in the root directory and
ignored by git.

This way when the src/server/CREDENTIALS file is tracked (and deleted) a
copy of the CREDENTIALS are still maintained in the root directory.

Just need to set up the build script to replace the missing CREDENTIALS
file with the copy before building if it is found to be missing.